moved article related functions into articles.js
